いろいろちまちま改良

+++ sources/Text.hpp
@@ -32,7 +32,7 @@ public:
         _textLayout.Reset();
         Graphics::GetTextFactory().CreateTextLayout(text.c_str(), (UINT32)text.length(), textFormat.Get(), textMetrics.width, textMetrics.height, _textLayout.GetAddressOf());
 
-        std::unique_ptr<BYTE[]> buffer(new BYTE[(int)textMetrics.width * (int)textMetrics.height * 4]);
+        std::unique_ptr<BYTE[]> buffer(new BYTE[int(textMetrics.width * textMetrics.height) * 4]);
         _texture.Create(buffer.get(), (int)textMetrics.width, (int)textMetrics.height);
 
         ComPtr<IDXGISurface> surface = nullptr;
